move registry_check
